tests: Dedicated file for Descriptor Indexing
