+++ tests/unit/descriptors.cpp
@@ -2915,415 +2915,6 @@ TEST_F(NegativeDescriptors, PushDescriptorSetLayoutWithoutExtension) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeDescriptors, DescriptorIndexingSetLayoutWithoutExtension) {
-    TEST_DESCRIPTION("Create an update_after_bind set layout without loading the needed extension.");
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT;
-
-    VkDescriptorSetLayout ds_layout = VK_NULL_HANDLE;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "UNASSIGNED-CoreValidation-DrawState-ExtensionNotEnabled");
-    vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &ds_layout);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeDescriptors, DescriptorIndexingSetLayout) {
-    TEST_DESCRIPTION("Exercise various create/allocate-time errors related to VK_EXT_descriptor_indexing.");
-
-    AddRequiredExtensions(VK_EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    // Create a device that enables all supported indexing features except descriptorBindingUniformBufferUpdateAfterBind
-    auto indexing_features = LvlInitStruct<VkPhysicalDeviceDescriptorIndexingFeaturesEXT>();
-    auto features2 = GetPhysicalDeviceFeatures2(indexing_features);
-
-    indexing_features.descriptorBindingUniformBufferUpdateAfterBind = VK_FALSE;
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    std::array<VkDescriptorBindingFlagsEXT, 2> flags = {
-        {VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT, VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT}};
-    auto flags_create_info = LvlInitStruct<VkDescriptorSetLayoutBindingFlagsCreateInfoEXT>();
-    flags_create_info.bindingCount = size32(flags);
-    flags_create_info.pBindingFlags = nullptr;
-
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>(&flags_create_info);
-    ds_layout_ci.bindingCount = 1;
-    ds_layout_ci.pBindings = &binding;
-
-    VkDescriptorSetLayout dsl = VK_NULL_HANDLE;
-    {
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutBindingFlagsCreateInfo-pBindingFlags-parameter");
-        vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &dsl);
-        m_errorMonitor->VerifyFound();
-    }
-
-    {
-        // VU for VkDescriptorSetLayoutBindingFlagsCreateInfoEXT::bindingCount
-        flags_create_info.pBindingFlags = flags.data();
-        flags_create_info.bindingCount = 2;
-
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutBindingFlagsCreateInfo-bindingCount-03002");
-        vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &dsl);
-        m_errorMonitor->VerifyFound();
-    }
-
-    flags_create_info.bindingCount = 1;
-
-    {
-        // set is missing UPDATE_AFTER_BIND_POOL flag.
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-flags-03000");
-        // binding uses a feature we disabled
-        m_errorMonitor->SetDesiredFailureMsg(
-            kErrorBit, "VUID-VkDescriptorSetLayoutBindingFlagsCreateInfo-descriptorBindingUniformBufferUpdateAfterBind-03005");
-        vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &dsl);
-        m_errorMonitor->VerifyFound();
-    }
-    {
-        ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT;
-        ds_layout_ci.bindingCount = 0;
-        flags_create_info.bindingCount = 0;
-        vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-        VkDescriptorPoolSize pool_size = {binding.descriptorType, binding.descriptorCount};
-        auto dspci = LvlInitStruct<VkDescriptorPoolCreateInfo>();
-        dspci.poolSizeCount = 1;
-        dspci.pPoolSizes = &pool_size;
-        dspci.maxSets = 1;
-        vk_testing::DescriptorPool pool(*m_device, dspci);
-
-        auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>();
-        ds_alloc_info.descriptorPool = pool.handle();
-        ds_alloc_info.descriptorSetCount = 1;
-        ds_alloc_info.pSetLayouts = &ds_layout.handle();
-
-        VkDescriptorSet ds = VK_NULL_HANDLE;
-        // mismatch between descriptor set and pool
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetAllocateInfo-pSetLayouts-03044");
-        vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
-        m_errorMonitor->VerifyFound();
-    }
-
-    if (indexing_features.descriptorBindingVariableDescriptorCount) {
-        VkDescriptorPoolSize pool_size = {binding.descriptorType, 3};
-        auto dspci = LvlInitStruct<VkDescriptorPoolCreateInfo>();
-        dspci.poolSizeCount = 1;
-        dspci.pPoolSizes = &pool_size;
-        dspci.maxSets = 2;
-        vk_testing::DescriptorPool pool(*m_device, dspci);
-        {
-            ds_layout_ci.flags = 0;
-            ds_layout_ci.bindingCount = 1;
-            flags_create_info.bindingCount = 1;
-            flags[0] = VK_DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT;
-            vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-            auto count_alloc_info = LvlInitStruct<VkDescriptorSetVariableDescriptorCountAllocateInfoEXT>();
-            count_alloc_info.descriptorSetCount = 1;
-            // Set variable count larger than what was in the descriptor binding
-            uint32_t variable_count = 2;
-            count_alloc_info.pDescriptorCounts = &variable_count;
-
-            auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>(&count_alloc_info);
-            ds_alloc_info.descriptorPool = pool.handle();
-            ds_alloc_info.descriptorSetCount = 1;
-            ds_alloc_info.pSetLayouts = &ds_layout.handle();
-
-            VkDescriptorSet ds = VK_NULL_HANDLE;
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit,
-                                                 "VUID-VkDescriptorSetVariableDescriptorCountAllocateInfo-pSetLayouts-03046");
-            vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
-            m_errorMonitor->VerifyFound();
-        }
-        {
-            // Now update descriptor set with a size that falls within the descriptor set layout size but that is more than the
-            // descriptor set size
-            binding.descriptorCount = 3;
-            vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-            auto count_alloc_info = LvlInitStruct<VkDescriptorSetVariableDescriptorCountAllocateInfoEXT>();
-            count_alloc_info.descriptorSetCount = 1;
-            uint32_t variable_count = 2;
-            count_alloc_info.pDescriptorCounts = &variable_count;
-
-            auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>(&count_alloc_info);
-            ds_alloc_info.descriptorPool = pool.handle();
-            ds_alloc_info.descriptorSetCount = 1;
-            ds_alloc_info.pSetLayouts = &ds_layout.handle();
-
-            VkDescriptorSet ds;
-            VkResult err = vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
-            ASSERT_VK_SUCCESS(err);
-            VkBufferObj buffer;
-            VkMemoryPropertyFlags reqs = 0;
-            buffer.init_as_dst(*m_device, 128 * 128, reqs);
-            VkDescriptorBufferInfo buffer_info[3] = {};
-            for (int i = 0; i < 3; i++) {
-                buffer_info[i].buffer = buffer.handle();
-                buffer_info[i].offset = 0;
-                buffer_info[i].range = 128 * 128;
-            }
-            VkWriteDescriptorSet descriptor_writes[1] = {};
-            descriptor_writes[0] = LvlInitStruct<VkWriteDescriptorSet>();
-            descriptor_writes[0].dstSet = ds;
-            descriptor_writes[0].dstBinding = 0;
-            descriptor_writes[0].descriptorCount = 3;
-            descriptor_writes[0].descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-            descriptor_writes[0].pBufferInfo = buffer_info;
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-dstArrayElement-00321");
-            vk::UpdateDescriptorSets(m_device->device(), 1, descriptor_writes, 0, NULL);
-            m_errorMonitor->VerifyFound();
-        }
-    }
-}
-
-TEST_F(NegativeDescriptors, DescriptorIndexingUpdateAfterBind) {
-    TEST_DESCRIPTION("Exercise errors for updating a descriptor set after it is bound.");
-
-    AddRequiredExtensions(VK_EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME);
-    AddRequiredExtensions(VK_KHR_MAINTENANCE_3_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    // Create a device that enables all supported indexing features except descriptorBindingUniformBufferUpdateAfterBind
-    auto indexing_features = LvlInitStruct<VkPhysicalDeviceDescriptorIndexingFeaturesEXT>();
-    auto features2 = GetPhysicalDeviceFeatures2(indexing_features);
-
-    indexing_features.descriptorBindingUniformBufferUpdateAfterBind = VK_FALSE;
-
-    if (VK_FALSE == indexing_features.descriptorBindingStorageBufferUpdateAfterBind) {
-        GTEST_SKIP() << "Test requires (unsupported) descriptorBindingStorageBufferUpdateAfterBind";
-    }
-    if (VK_FALSE == features2.features.fragmentStoresAndAtomics) {
-        GTEST_SKIP() << "Test requires (unsupported) fragmentStoresAndAtomics";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT));
-    ASSERT_NO_FATAL_FAILURE(InitViewport());
-    ASSERT_NO_FATAL_FAILURE(InitRenderTarget());
-
-    VkDescriptorBindingFlagsEXT flags[3] = {0, VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT,
-                                            VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT};
-    auto flags_create_info = LvlInitStruct<VkDescriptorSetLayoutBindingFlagsCreateInfoEXT>();
-    flags_create_info.bindingCount = 3;
-    flags_create_info.pBindingFlags = &flags[0];
-
-    // Descriptor set has two bindings - only the second is update_after_bind
-    VkDescriptorSetLayoutBinding binding[3] = {
-        {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-        {1, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-        {2, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-    };
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>(&flags_create_info);
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT;
-    ds_layout_ci.bindingCount = 3;
-    ds_layout_ci.pBindings = &binding[0];
-    vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-    VkDescriptorPoolSize pool_sizes[3] = {
-        {binding[0].descriptorType, binding[0].descriptorCount},
-        {binding[1].descriptorType, binding[1].descriptorCount},
-        {binding[2].descriptorType, binding[2].descriptorCount},
-    };
-    auto dspci = LvlInitStruct<VkDescriptorPoolCreateInfo>();
-    dspci.flags = VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT_EXT;
-    dspci.poolSizeCount = 3;
-    dspci.pPoolSizes = &pool_sizes[0];
-    dspci.maxSets = 1;
-    vk_testing::DescriptorPool pool(*m_device, dspci);
-
-    auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>();
-    ds_alloc_info.descriptorPool = pool.handle();
-    ds_alloc_info.descriptorSetCount = 1;
-    ds_alloc_info.pSetLayouts = &ds_layout.handle();
-
-    VkDescriptorSet ds = VK_NULL_HANDLE;
-    VkResult err = vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
-    ASSERT_VK_SUCCESS(err);
-
-    auto buffCI = LvlInitStruct<VkBufferCreateInfo>();
-    buffCI.size = 1024;
-    buffCI.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT | VK_BUFFER_USAGE_STORAGE_BUFFER_BIT;
-
-    vk_testing::Buffer dynamic_uniform_buffer(*m_device, buffCI, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT);
-
-    VkDescriptorBufferInfo buffInfo[2] = {};
-    buffInfo[0].buffer = dynamic_uniform_buffer.handle();
-    buffInfo[0].offset = 0;
-    buffInfo[0].range = 1024;
-
-    VkWriteDescriptorSet descriptor_write[2] = {};
-    descriptor_write[0] = LvlInitStruct<VkWriteDescriptorSet>();
-    descriptor_write[0].dstSet = ds;
-    descriptor_write[0].dstBinding = 0;
-    descriptor_write[0].descriptorCount = 1;
-    descriptor_write[0].descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    descriptor_write[0].pBufferInfo = buffInfo;
-    descriptor_write[1] = descriptor_write[0];
-    descriptor_write[1].dstBinding = 1;
-    descriptor_write[1].descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-
-    auto pipeline_layout_ci = LvlInitStruct<VkPipelineLayoutCreateInfo>();
-    pipeline_layout_ci.setLayoutCount = 1;
-    pipeline_layout_ci.pSetLayouts = &ds_layout.handle();
-
-    vk_testing::PipelineLayout pipeline_layout(*m_device, pipeline_layout_ci);
-
-    // Create a dummy pipeline, since VL inspects which bindings are actually used at draw time
-    char const *fsSource = R"glsl(
-        #version 450
-        layout(location=0) out vec4 color;
-        layout(set=0, binding=0) uniform foo0 { float x0; } bar0;
-        layout(set=0, binding=1) buffer  foo1 { float x1; } bar1;
-        layout(set=0, binding=2) buffer  foo2 { float x2; } bar2;
-        void main(){
-           color = vec4(bar0.x0 + bar1.x1 + bar2.x2);
-        }
-    )glsl";
-
-    VkShaderObj vs(this, kVertexMinimalGlsl, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    VkPipelineObj pipe(m_device);
-    pipe.SetViewport(m_viewports);
-    pipe.SetScissor(m_scissors);
-    pipe.AddDefaultColorAttachment();
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.CreateVKPipeline(pipeline_layout.handle(), m_renderPass);
-
-    // Make both bindings valid before binding to the command buffer
-    vk::UpdateDescriptorSets(m_device->device(), 2, &descriptor_write[0], 0, NULL);
-
-    auto submit_info = LvlInitStruct<VkSubmitInfo>();
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_commandBuffer->handle();
-
-    // Two subtests. First only updates the update_after_bind binding and expects
-    // no error. Second updates the other binding and expects an error when the
-    // command buffer is ended.
-    for (uint32_t i = 0; i < 2; ++i) {
-        m_commandBuffer->begin();
-
-        vk::CmdBindDescriptorSets(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1, &ds,
-                                  0, NULL);
-
-        m_commandBuffer->BeginRenderPass(m_renderPassBeginInfo);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-        vk::CmdDraw(m_commandBuffer->handle(), 0, 0, 0, 0);
-        vk::CmdEndRenderPass(m_commandBuffer->handle());
-
-        // Valid to update binding 1 after being bound
-        vk::UpdateDescriptorSets(m_device->device(), 1, &descriptor_write[1], 0, NULL);
-
-        if (i == 0) {
-            // expect no errors
-            m_commandBuffer->end();
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdDraw-None-08114");
-            vk::QueueSubmit(m_device->m_queue, 1, &submit_info, VK_NULL_HANDLE);
-            m_errorMonitor->VerifyFound();
-            vk::QueueWaitIdle(m_device->m_queue);
-        } else {
-            // Invalid to update binding 0 after being bound. But the error is actually
-            // generated during vk::EndCommandBuffer
-            vk::UpdateDescriptorSets(m_device->device(), 1, &descriptor_write[0], 0, NULL);
-
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit,
-                                                 "UNASSIGNED-CoreValidation-DrawState-InvalidCommandBuffer-VkDescriptorSet");
-
-            vk::EndCommandBuffer(m_commandBuffer->handle());
-            m_errorMonitor->VerifyFound();
-        }
-    }
-}
-
-TEST_F(NegativeDescriptors, DescriptorIndexingSetNonIdenticalWrite) {
-    TEST_DESCRIPTION("VkWriteDescriptorSet must have identical VkDescriptorBindingFlagBits");
-
-    AddRequiredExtensions(VK_EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto indexing_features = LvlInitStruct<VkPhysicalDeviceDescriptorIndexingFeatures>();
-    auto features2 = GetPhysicalDeviceFeatures2(indexing_features);
-
-    if (VK_FALSE == indexing_features.descriptorBindingStorageBufferUpdateAfterBind) {
-        GTEST_SKIP() << "Test requires (unsupported) descriptorBindingStorageBufferUpdateAfterBind";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    // not all identical VkDescriptorBindingFlags flags
-    VkDescriptorBindingFlags flags[3] = {VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT, 0,
-                                         VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT};
-    auto flags_create_info = LvlInitStruct<VkDescriptorSetLayoutBindingFlagsCreateInfo>();
-    flags_create_info.bindingCount = 3;
-    flags_create_info.pBindingFlags = &flags[0];
-
-    VkDescriptorSetLayoutBinding binding[3] = {
-        {0, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-        {1, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-        {2, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr},
-    };
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>(&flags_create_info);
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT;
-    ds_layout_ci.bindingCount = 3;
-    ds_layout_ci.pBindings = &binding[0];
-    vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-    VkDescriptorPoolSize pool_sizes = {VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 3};
-    auto dspci = LvlInitStruct<VkDescriptorPoolCreateInfo>();
-    dspci.flags = VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT;
-    dspci.poolSizeCount = 1;
-    dspci.pPoolSizes = &pool_sizes;
-    dspci.maxSets = 3;
-    vk_testing::DescriptorPool pool(*m_device, dspci);
-
-    auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>();
-    ds_alloc_info.descriptorPool = pool.handle();
-    ds_alloc_info.descriptorSetCount = 1;
-    ds_alloc_info.pSetLayouts = &ds_layout.handle();
-    VkDescriptorSet ds = VK_NULL_HANDLE;
-    ASSERT_VK_SUCCESS(vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds));
-
-    auto buff_create_info = LvlInitStruct<VkBufferCreateInfo>();
-    buff_create_info.size = 1024;
-    buff_create_info.usage = VK_BUFFER_USAGE_STORAGE_BUFFER_BIT;
-    VkBufferObj buffer;
-    buffer.init(*m_device, buff_create_info);
-
-    VkDescriptorBufferInfo bufferInfo[3] = {};
-    bufferInfo[0].buffer = buffer.handle();
-    bufferInfo[0].offset = 0;
-    bufferInfo[0].range = 1024;
-    bufferInfo[1] = bufferInfo[0];
-    bufferInfo[2] = bufferInfo[0];
-
-    auto descriptor_write = LvlInitStruct<VkWriteDescriptorSet>();
-    descriptor_write.dstSet = ds;
-    descriptor_write.dstBinding = 0;
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_write.pBufferInfo = bufferInfo;
-    // If the dstBinding has fewer than descriptorCount, remainder will be used to update the subsequent binding
-    descriptor_write.descriptorCount = 3;
-
-    // binding 1 has a different VkDescriptorBindingFlags
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-dstArrayElement-00321");
-    vk::UpdateDescriptorSets(m_device->device(), 1, &descriptor_write, 0, NULL);
-    m_errorMonitor->VerifyFound();
-}
-
 TEST_F(NegativeDescriptors, AllocatePushDescriptorSet) {
     TEST_DESCRIPTION("Attempt to allocate a push descriptor set.");
     AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
@@ -5333,56 +4924,6 @@ TEST_F(NegativeDescriptors, AllocatingVariableDescriptorSets) {
     VkDescriptorSet ds;
     VkResult err = vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
     ASSERT_VK_SUCCESS(err);
-}
-
-TEST_F(NegativeDescriptors, DescriptorSetLayoutBindings) {
-    TEST_DESCRIPTION("Create descriptor set layout with incompatible bindings.");
-
-    AddRequiredExtensions(VK_EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    auto indexing_features = LvlInitStruct<VkPhysicalDeviceDescriptorIndexingFeaturesEXT>();
-    auto features2 = GetPhysicalDeviceFeatures2(indexing_features);
-    if (VK_FALSE == indexing_features.descriptorBindingUniformBufferUpdateAfterBind) {
-        GTEST_SKIP() << "Test requires (unsupported) descriptorBindingStorageBufferUpdateAfterBind";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    VkDescriptorSetLayoutBinding update_binding = {};
-    update_binding.binding = 0;
-    update_binding.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    update_binding.descriptorCount = 1;
-    update_binding.stageFlags = VK_SHADER_STAGE_ALL;
-    update_binding.pImmutableSamplers = nullptr;
-
-    VkDescriptorSetLayoutBinding dynamic_binding = {};
-    dynamic_binding.binding = 1;
-    dynamic_binding.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC;
-    dynamic_binding.descriptorCount = 1;
-    dynamic_binding.stageFlags = VK_SHADER_STAGE_ALL;
-    dynamic_binding.pImmutableSamplers = nullptr;
-
-    VkDescriptorSetLayoutBinding bindings[2] = {update_binding, dynamic_binding};
-
-    VkDescriptorBindingFlags flags[2] = {VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT, 0};
-
-    auto flags_create_info = LvlInitStruct<VkDescriptorSetLayoutBindingFlagsCreateInfoEXT>();
-    flags_create_info.bindingCount = 2;
-    flags_create_info.pBindingFlags = flags;
-
-    VkDescriptorSetLayoutCreateInfo create_info = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    create_info.pNext = &flags_create_info;
-    create_info.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT;
-    create_info.bindingCount = 2;
-    create_info.pBindings = bindings;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-descriptorType-03001");
-    VkDescriptorSetLayout setLayout;
-    vk::CreateDescriptorSetLayout(m_device->handle(), &create_info, nullptr, &setLayout);
-    m_errorMonitor->VerifyFound();
 }
 
 TEST_F(NegativeDescriptors, DescriptorSetLayoutBinding) {
